cbkb-ASCI/requirements.txt
Pillow==8.3.2
tqdm==4.62.3
